Switch to using LZ4 compression by default
Versions of Waypipe >= 0.8.2 always send the compression type in
ssh mode. When using <0.8.2 versions of Waypipe to connect to
versions following this commit, one should use either the -c lz4
or the -c zstd option when in ssh mode. Waypipe >= 0.8.2 will also
provide an error if the client and server use different
compression types.